requirements.txt
torch
Pillow
loguru
numpy
argparse
matplotlib
torchvision
